chal/requirements.txt
rsa==4.9
pycryptodome==3.15.0